HTBHF-629 refactor check details route POST handler
