+++ src/web/routes/application/check/post.js
@@ -1,40 +1,13 @@
 const httpStatus = require('http-status-codes')
 const request = require('request-promise')
 
-const { toDateString, wrapError } = require('../common/formatters')
-const { YES } = require('../common/constants')
+const { wrapError } = require('../common/formatters')
 const { logger } = require('../../../logger')
 const { REQUEST_ID_HEADER } = require('../../../server/headers')
 const { stateMachine, states, actions } = require('../common/state-machine')
+const { createRequestBody } = require('./create-request-body')
 
 const CLAIMS_ENDPOINT = `/v1/claims`
-
-const createExpectedDeliveryDate = (claim) => {
-  if (claim.areYouPregnant === YES) {
-    return toDateString(
-      claim['expectedDeliveryDate-day'],
-      claim['expectedDeliveryDate-month'],
-      claim['expectedDeliveryDate-year']
-    )
-  }
-  return null
-}
-
-const createRequestBody = (claim) => {
-  return {
-    firstName: claim.firstName,
-    lastName: claim.lastName,
-    nino: claim.nino,
-    dateOfBirth: claim.dateOfBirth,
-    cardDeliveryAddress: {
-      addressLine1: claim.addressLine1,
-      addressLine2: claim.addressLine2,
-      townOrCity: claim.townOrCity,
-      postcode: claim.postcode
-    },
-    expectedDeliveryDate: createExpectedDeliveryDate(claim)
-  }
-}
 
 const postCheck = (steps, config) => async (req, res, next) => {
   try {
@@ -65,6 +38,5 @@ const postCheck = (steps, config) => async (req, res, next) => {
 }
 
 module.exports = {
-  postCheck,
-  createRequestBody
+  postCheck
 }
